+++ NEWS.md
@@ -1,13 +1,18 @@
 # tensorflow (development version)
 
-- `as_tensor()` now converts atomic (non-array) R double vectors to "float64" dtype.
-  Previously they would in some circumstances get converted to "float32".
+- `as_tensor()` now coerces bare R atomic vectors to R arrays before conversion.
+  As a consequence, by default, R atomic double vectors now coerce to 
+  'float64' dtype tensors instead of 'float32'.
+  
+- `shape()` gains the ability to accept vectors of length > 1 in `...`, 
+  including other `tf.TensorShape`s. Shapes are automatically flattened.
+
 # tensorflow 2.9.0
 
 - Generic method updates:
   - New methods:
       all(), any(), sum(), prod(), min(), max(), mean(), range(),
-      cbind(), rbind(), t(), aperm(), sort(),
+      cbind(), rbind(), t(), aperm(), sort(), 
       as.vector(), as.character(), as.raster(),
       is.infinite(), is.finite(), is.nan()
   - `^` will now invoke `tf.square()` or `tf.sqrt()` directly when appropriate